refactor(TotalRound): 입력받는 시도할 횟수도 총 라운드라는 클래스로 구현
도메인을 객체지향적으로 설계했을 때 총 라운드 수라는 핵심 개념을 추출할 수 있었음
따라서 클래스를 만들고 객체를 생성하여 사용하도록 구현
