+++ src/main/java/views/InputView.java
@@ -10,10 +10,7 @@ public class InputView {
     private static final String SEPARATOR = ",";
     private static final String INPUT_ROUND_NUMBER_MESSAGE = "시도할 회수는 몇회인가요?";
     private static final String EMPTY_EXCEPTION = "[ERROR] 공백을 입력할 수 없습니다.";
-    private static final String NOT_NUMBER_EXCEPTION = "[ERROR] 숫자를 입력해야 합니다.";
-    private static final String IMPOSSIBLE_NUMBER_ZERO_EXCEPTION = "[ERROR] 0은 임력할 수 없습니다.";
     private static final String EMPTY_INPUT = "";
-    private static final String ZERO = "0";
 
     public static List<String> inputCarNames(Scanner scanner) {
         System.out.println(INPUT_CAR_NAME_MESSAGE);
@@ -23,32 +20,16 @@ public class InputView {
                 .collect(Collectors.toList());
     }
 
-    public static int inputRoundNumber(Scanner scanner) {
+    public static String inputRoundNumber(Scanner scanner) {
         System.out.println(INPUT_ROUND_NUMBER_MESSAGE);
         String roundNumber = scanner.nextLine();
         isEmptyValue(roundNumber);
-        isNumber(roundNumber);
-        isZero(roundNumber);
-        return Integer.parseInt(roundNumber);
+        return roundNumber;
     }
 
     private static void isEmptyValue(String input) {
         if (input.equals(EMPTY_INPUT)) {
             throw new IllegalArgumentException(EMPTY_EXCEPTION);
-        }
-    }
-
-    private static void isNumber(String input) {
-        try {
-            Integer.parseInt(input);
-        } catch (Exception e) {
-            throw new IllegalArgumentException(NOT_NUMBER_EXCEPTION);
-        }
-    }
-
-    private static void isZero(String input) {
-        if (input.equals(ZERO)) {
-            throw new IllegalArgumentException(IMPOSSIBLE_NUMBER_ZERO_EXCEPTION);
         }
     }
 }
